PIC!6F8x API update. Adds the "makefile_flags" option to .TOML file. Now it is possible to chose the devices from settings.json in vscode.

+++ ports/Microchip/PIC16F-generic/api/pin/read.c.v
@@ -10,6 +10,6 @@ module pin
 @[inline]
 pub fn read(name u8) u8 {
 	unsafe {
-		return (*(&C.PORTA + (name >> 3))  >>  (name - ((name >> 3) << 3)))  &  0x01
-	}// return (*(&C.PORTA + (name / 8))   >>  (name % 8))  &  0x01
+		return	u8( (*(&C.PORTA + (name >> 3))  >>  (name - ((name >> 3) << 3)))  &  0x01 )
+	}// return	u8( (*(&C.PORTA + (name / 8))   >>  (name % 8))  &  0x01 )
 }
